Improve statistics for very large reached sets. Mostly we now do not calculcate some details about large reached sets because this would be very expensive (taking minutes of time).
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@8790 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/core/algorithm/CPAAlgorithm.java
@@ -74,7 +74,7 @@ public class CPAAlgorithm implements Algorithm, StatisticsProvider {
 
     private int   countIterations   = 0;
     private int   maxWaitlistSize   = 0;
-    private int   countWaitlistSize = 0;
+    private long  countWaitlistSize = 0;
     private int   countSuccessors   = 0;
     private int   maxSuccessors     = 0;
     private int   countMerge        = 0;
